requirements.txt
pytorch==1.10.0
pytorch-lightning==1.5.8
torchvision==0.6.2
pillow==8.3.2
hydra-core==1.1.1
pandas==1.3.4
opencv==4.5.5
simple-ITK
scikit-learn==1.0.2
selective_search
SimpleITK

